+++ R/tidy_model_expressions.R
@@ -27,57 +27,36 @@ tidy_model_expressions <- function(data,
                                    effsize.type = "omega",
                                    ...) {
 
-  # standardize
+  # standardize the statistic naming
   statistic <- substring(tolower(statistic), 1L, 1L)
 
-  # all operations will have to be done rowwise
-  data %<>% rowwise()
+  # convert the necessary columns to character type for expression
+  df <- .data_to_char(data, k)
+
+  # effect size text for the expression (common for t, z, and chi^2)
+  es.text <- list(quote(widehat(italic(beta))))
 
   # t-statistic --------------------------------
 
   if (statistic == "t") {
-    data %<>%
-      mutate(
-        label = case_when(
-          is.na(df.error) || is.infinite(df.error) ~ paste0(
-            "list(~widehat(italic(beta))=='", format_value(estimate, k),
-            "', ~italic(t)=='", format_value(statistic, k),
-            "', ~italic(p)=='", format_num(p.value, k, TRUE), "')"
-          ),
-          TRUE ~ paste0(
-            "list(~widehat(italic(beta))=='", format_value(estimate, k),
-            "', ~italic(t)", "('", format_value(df.error, 0L), "')=='", format_value(statistic, k),
-            "', ~italic(p)=='", format_num(p.value, k, TRUE), "')"
-          )
-        )
+    df %<>% mutate(
+      label = case_when(
+        df.error %in% c("", "Inf") ~ glue("list({es.text}=='{estimate}', italic(t)=='{statistic}', italic(p)=='{p.value}')"),
+        TRUE ~ glue("list({es.text}=='{estimate}', italic(t)('{df.error}')=='{statistic}', italic(p)=='{p.value}')")
       )
+    )
   }
 
   # z-statistic ---------------------------------
 
-  # if the statistic is z-value
   if (statistic == "z") {
-    data %<>%
-      mutate(
-        label = paste0(
-          "list(~widehat(italic(beta))=='", format_value(estimate, k),
-          "', ~italic(z)=='", format_value(statistic, k),
-          "', ~italic(p)=='", format_num(p.value, k, TRUE), "')"
-        )
-      )
+    df %<>% mutate(label = glue("list({es.text}=='{estimate}', italic(z)=='{statistic}', italic(p)=='{p.value}')"))
   }
 
   # chi^2-statistic -----------------------------
 
   if (statistic == "c") {
-    data %<>%
-      mutate(
-        label = paste0(
-          "list(~widehat(italic(beta))=='", format_value(estimate, k),
-          "', ~italic(chi)^2~", "('", format_value(df.error, 0L), "')=='", format_value(statistic, k),
-          "', ~italic(p)=='", format_num(p.value, k, TRUE), "')"
-        )
-      )
+    df %<>% mutate(label = glue("list({es.text}=='{estimate}', italic(chi)^2*('{df.error}')=='{statistic}', italic(p)=='{p.value}')"))
   }
 
   # f-statistic ---------------------------------
@@ -88,17 +67,9 @@ tidy_model_expressions <- function(data,
     if (effsize.type == "omega") es.text <- list(quote(widehat(italic(omega)[p]^2)))
 
     # which effect size is needed?
-    data %<>%
-      mutate(
-        label = paste0(
-          "list(~italic(F)", "('", format_value(df, 0L), "'*\",\"*'", format_value(df.error, 0L),
-          "')=='", format_value(statistic, k),
-          "', ~italic(p)=='", format_num(p.value, k, TRUE),
-          "', ~", es.text, "=='", format_value(estimate, k), "')"
-        )
-      )
+    df %<>% mutate(label = glue("list({es.text}=='{estimate}', italic(F)('{df}', '{df.error}')=='{statistic}', italic(p)=='{p.value}')"))
   }
 
-  # return the ungrouped dataframe
-  ungroup(data)
+  # add the label column to the original dataframe
+  left_join(data, select(df, term, label), by = "term")
 }