+++ tests/testthat/test-contingency_table.R
@@ -18,7 +18,7 @@ test_that(
     # testing all details
     set.seed(123)
     expect_snapshot(select(df1, -expression))
-    expect_snapshot(df1$expression[[1]])
+    expect_snapshot(as.character(df1$expression[[1]]))
 
     # with counts
     set.seed(123)
@@ -32,7 +32,7 @@ test_that(
     # testing all details
     set.seed(123)
     expect_snapshot(select(df2, -expression))
-    expect_snapshot(df2$expression[[1]])
+    expect_snapshot(as.character(df2$expression[[1]]))
 
     # contingency tab - with NAs --------------------------------------
 
@@ -48,7 +48,7 @@ test_that(
     # testing all details
     set.seed(123)
     expect_snapshot(select(df3, -expression))
-    expect_snapshot(df3$expression[[1]])
+    expect_snapshot(as.character(df3$expression[[1]]))
   }
 )
 
@@ -94,7 +94,7 @@ test_that(
     # testing all details
     set.seed(123)
     expect_snapshot(select(df1, -expression))
-    expect_snapshot(df1$expression[[1]])
+    expect_snapshot(as.character(df1$expression[[1]]))
 
     # paired data with NAs  ---------------------------------------------
 
@@ -126,7 +126,7 @@ test_that(
     # testing all details
     set.seed(123)
     expect_snapshot(select(df2, -expression))
-    expect_snapshot(df2$expression[[1]])
+    expect_snapshot(as.character(df2$expression[[1]]))
   }
 )
 
@@ -147,7 +147,7 @@ test_that(
     # testing all details
     set.seed(123)
     expect_snapshot(select(df1, -expression))
-    expect_snapshot(df1$expression[[1]])
+    expect_snapshot(as.character(df1$expression[[1]]))
 
     # with counts
     set.seed(123)
@@ -160,7 +160,7 @@ test_that(
     # testing all details
     set.seed(123)
     expect_snapshot(select(df2, -expression))
-    expect_snapshot(df2$expression[[1]])
+    expect_snapshot(as.character(df2$expression[[1]]))
 
     # one-sample test (with NAs) -------------------------------------
 
@@ -175,7 +175,7 @@ test_that(
     # testing all details
     set.seed(123)
     expect_snapshot(select(df3, -expression))
-    expect_snapshot(df3$expression[[1]])
+    expect_snapshot(as.character(df3$expression[[1]]))
   }
 )
 
@@ -192,8 +192,8 @@ test_that(
       type = "bayes"
     )
 
-    # check bayes factor values
-    expect_equal(df1$bf10, 0.2465787, tolerance = 0.001)
+    expect_snapshot(select(df1, -expression))
+    expect_snapshot(as.character(df1$expression[[1]]))
 
     # expr
     set.seed(123)
@@ -205,7 +205,8 @@ test_that(
       top.text = "duh"
     )
 
-    expect_snapshot(df2$expression[[1]])
+    expect_snapshot(select(df2, -expression))
+    expect_snapshot(as.character(df2$expression[[1]]))
   }
 )
 
@@ -272,10 +273,9 @@ test_that(
 
     # expr text
     expect_snapshot(list(
-      expr_text1$expression[[1]],
-      expr_text2$expression[[1]],
-      expr_text3$expression[[1]]
+      as.character(expr_text1$expression[[1]]),
+      as.character(expr_text2$expression[[1]]),
+      as.character(expr_text3$expression[[1]])
     ))
   }
 )
-