+++ tests/testthat/test-tidy_model_expressions.R
@@ -9,10 +9,31 @@ test_that("tidy_model_expressions works", {
   mod_t <- lm(wt ~ mpg, data = mtcars)
 
   set.seed(123)
-  expect_snapshot(suppressWarnings(tidy_model_expressions(
+  df_t <- suppressWarnings(tidy_model_expressions(
     tidy_model_parameters(mod_t),
     statistic = "t"
-  )))
+  ))
+
+  expect_snapshot(select(df_t, -label))
+  expect_snapshot(df_t$label)
+
+  # with NA df.error
+  set.seed(123)
+  df_t_na <- suppressWarnings(tidy_model_expressions(
+    mutate(tidy_model_parameters(mod_t), df.error = NA_real_),
+    statistic = "t"
+  ))
+
+  expect_snapshot(df_t_na$label)
+
+  # with infinity as error
+  set.seed(123)
+  df_t_inf <- suppressWarnings(tidy_model_expressions(
+    mutate(tidy_model_parameters(mod_t), df.error = Inf),
+    statistic = "t"
+  ))
+
+  expect_snapshot(df_t_inf$label)
 
   ## chi2-statistic --------------------------------
 
@@ -26,10 +47,13 @@ test_that("tidy_model_expressions works", {
   )
 
   set.seed(123)
-  expect_snapshot(suppressWarnings(tidy_model_expressions(
+  df_chi <- suppressWarnings(tidy_model_expressions(
     tidy_model_parameters(mod_chi),
     statistic = "chi"
-  )))
+  ))
+
+  expect_snapshot(select(df_chi, -label))
+  expect_snapshot(df_chi$label)
 
   ## z-statistic --------------------------------
 
@@ -45,8 +69,11 @@ test_that("tidy_model_expressions works", {
   )
 
   set.seed(123)
-  expect_snapshot(suppressWarnings(tidy_model_expressions(
+  df_z <- suppressWarnings(tidy_model_expressions(
     tidy_model_parameters(mod_z),
     statistic = "z"
-  )))
+  ))
+
+  expect_snapshot(select(df_z, -label))
+  expect_snapshot(df_z$label)
 })