+++ tests/testthat/test-tidy_model_expressions_F.R
@@ -3,27 +3,31 @@ test_that("tidy_model_expressions works - F", {
 
   ## F-statistic --------------------------------
 
-  if (packageVersion("parameters") >= "0.14.0.1") {
-    set.seed(123)
-    mod_f <- aov(yield ~ N * P + Error(block), npk)
+  set.seed(123)
+  mod_f <- aov(yield ~ N * P + Error(block), npk)
 
-    set.seed(123)
-    expect_snapshot(tidy_model_expressions(
-      tidy_model_parameters(mod_f,
-        omega_squared = "partial",
-        table_wide = TRUE
-      ),
-      statistic = "f"
-    ))
+  set.seed(123)
+  df1 <- tidy_model_expressions(
+    tidy_model_parameters(mod_f,
+      omega_squared = "partial",
+      table_wide = TRUE
+    ),
+    statistic = "f"
+  )
 
-    set.seed(123)
-    expect_snapshot(tidy_model_expressions(
-      tidy_model_parameters(mod_f,
-        eta_squared = "partial",
-        table_wide = TRUE
-      ),
-      statistic = "f",
-      effsize.type = "eta"
-    ))
-  }
+  expect_snapshot(select(df1, -label))
+  expect_snapshot(df1$label)
+
+  set.seed(123)
+  df2 <- tidy_model_expressions(
+    tidy_model_parameters(mod_f,
+      eta_squared = "partial",
+      table_wide = TRUE
+    ),
+    statistic = "f",
+    effsize.type = "eta"
+  )
+
+  expect_snapshot(select(df2, -label))
+  expect_snapshot(df2$label)
 })