+++ tests/testthat/test-two_sample_nonparametric.R
@@ -20,7 +20,7 @@ test_that(
     # testing all details
     set.seed(123)
     expect_snapshot(select(df, -expression))
-    expect_snapshot(df$expression[[1]])
+    expect_snapshot(as.character(df$expression[[1]]))
   }
 )
 
@@ -77,7 +77,7 @@ test_that(
     # testing all details
     set.seed(123)
     expect_snapshot(select(df2, -expression))
-    expect_snapshot(df2$expression[[1]])
+    expect_snapshot(as.character(df2$expression[[1]]))
   }
 )
 
@@ -139,6 +139,6 @@ test_that(
         paired = TRUE
       )
 
-    expect_equal(expr1, expr2)
+    expect_equal(expr1, expr2, ignore_attr = TRUE)
   }
 )